Adding scrape result / protocol
