--- hooply/market/__init__.py
@@ -1,10 +1,0 @@
-from peewee import *
-
-db = MySQLDatabase(
-    database="hooply",
-    host="localhost",
-    port=3306,
-    user="root",
-    password="test",
-    # pragmas={"journal_mode": "wal", "cache_size": 10000, "foreign_keys": 1},
-)
