+++ hooply/market/pipeline/pipeline.py
@@ -1,0 +1,5 @@
+def init_app():
+    pass
+    # Initialize redis queue (if fresh is specified)
+    #                         -> queue tasks
+    #                          -> schedule periodic jobs
